# solution/springboot-servers/SpringBoot-Server/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/themes/getThemes": {
      "post": {
        "tags": [
          "themes-controller"
        ],
        "operationId": "getThemesUserRequested",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListWrapperDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/getLatestReleasesMovies": {
      "post": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getLatestReleasesMovies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieCategoryDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/getBestMoviesOfAllTime": {
      "post": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getBestOfAllTimeMovies",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/themes/getPreciseTheme": {
      "get": {
        "tags": [
          "themes-controller"
        ],
        "operationId": "getPreciseTheme",
        "parameters": [
          {
            "name": "theme",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/studios/find": {
      "get": {
        "tags": [
          "studios-controller"
        ],
        "operationId": "getMoviesFromStudios",
        "parameters": [
          {
            "name": "studio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/top20ByRating": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getTop20ByRating",
        "parameters": [
          {
            "name": "yearOfRelease",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieCategoryDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/single-movie": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getMovieById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedMovieDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/getHomepageMovies": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getHomeMovies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HomepageDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/getAllTvSeries": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getAllTvSeries",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/movie/find": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getSearchedMovies",
        "parameters": [
          {
            "name": "movie",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/genre/find": {
      "get": {
        "tags": [
          "genres-controller"
        ],
        "operationId": "getMovieByGenre",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/crew/getDirector": {
      "get": {
        "tags": [
          "crew-controller"
        ],
        "operationId": "getDirector",
        "parameters": [
          {
            "name": "director",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/actors/getActorsByName": {
      "get": {
        "tags": [
          "actors-controller"
        ],
        "operationId": "getActorsByName",
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "movieId,asc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageMovieCategoryDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ListWrapperDTO": {
        "type": "object",
        "properties": {
          "moviesMatchedWithTheme": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieCategoryDTO"
            }
          }
        }
      },
      "MovieCategoryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "title": {
            "type": "string",
            "writeOnly": true
          },
          "poster": {
            "type": "string",
            "writeOnly": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "writeOnly": true
          },
          "posterLink": {
            "type": "string"
          },
          "movieTitle": {
            "type": "string"
          },
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "yearOfRelease": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageMovieCategoryDTO": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieCategoryDTO"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "nullHandling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignoreCase": {
            "type": "boolean"
          }
        }
      },
      "Actors": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "actorName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "Crew": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string"
          },
          "crewMemberName": {
            "type": "string"
          }
        }
      },
      "DetailedMovieDTO": {
        "type": "object",
        "properties": {
          "movie_id": {
            "type": "integer",
            "format": "int64"
          },
          "movie_title": {
            "type": "string"
          },
          "poster_link": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tagline": {
            "type": "string"
          },
          "year_of_release": {
            "type": "integer",
            "format": "int32"
          },
          "minutes": {
            "type": "number",
            "format": "float"
          },
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Actors"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genres"
            }
          },
          "themes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "directors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Crew"
            }
          },
          "oscars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Oscars"
            }
          },
          "correlated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieCategoryDTO"
            }
          },
          "rating": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Genres": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "genre": {
            "type": "string"
          }
        }
      },
      "Oscars": {
        "type": "object",
        "properties": {
          "yearOfRelease": {
            "type": "integer",
            "format": "int32"
          },
          "yearOfCeremony": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfCeremony": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string"
          },
          "candidateName": {
            "type": "string"
          },
          "movieTitle": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "hasWon": {
            "type": "boolean"
          }
        }
      },
      "CarouselDTO": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "tagline": {
            "type": "string"
          },
          "posterLink": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "writeOnly": true
          },
          "movieTitle": {
            "type": "string"
          },
          "movieId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "HomepageDTO": {
        "type": "object",
        "properties": {
          "carouselMovies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarouselDTO"
            }
          },
          "mostAwardedMovies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieCategoryDTO"
            }
          },
          "topRatedMovies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieCategoryDTO"
            }
          }
        }
      }
    }
  }
}